# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ImmoPro API",
        "description": "Documentation interactive de l'API ImmoPro — plateforme immobilière.\n\n## Authentification\nLa majorité des endpoints nécessitent un token Bearer (Laravel Sanctum).\nConnectez-vous via **POST /api/login** (Admin/Agent) ou **POST /api/client/login** (Client),\npuis cliquez sur le bouton **Authorize** et collez votre token.\n\n## Rôles\n- **admin** : accès complet à l'administration\n- **agent** : gestion des biens assignés et des visites\n- **client** : propriétaires / acheteurs / locataires",
        "contact": {
            "name": "Support ImmoPro",
            "email": "support@immopro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal de l'API ImmoPro"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentification Admin / Agent"
                ],
                "summary": "Connexion admin ou agent",
                "description": "Authentifie un utilisateur avec le rôle admin ou agent.",
                "operationId": "adminAgentLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@immopro.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MotDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "3|aB7cD9eF2gH4iJ6kL..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Compte bloqué ou suspendu"
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification Admin / Agent"
                ],
                "summary": "Déconnexion admin ou agent",
                "operationId": "adminAgentLogout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Authentification Admin / Agent"
                ],
                "summary": "Profil de l'utilisateur connecté (admin/agent)",
                "operationId": "adminAgentMe",
                "responses": {
                    "200": {
                        "description": "Profil récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "put": {
                "tags": [
                    "Authentification Admin / Agent"
                ],
                "summary": "Mise à jour du profil (admin/agent)",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmed",
                                        "maxLength": 100
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Diallo",
                                        "maxLength": 100
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmed.diallo@immopro.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+22501000000",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Côte d'Ivoire",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Abidjan",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil mis à jour."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/photo": {
            "post": {
                "tags": [
                    "Authentification Admin / Agent"
                ],
                "summary": "Changer la photo de profil (admin/agent)",
                "operationId": "updateProfilePhoto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "Image JPG, PNG ou WebP (max 5 Mo)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Photo de profil mise à jour."
                                        },
                                        "profile_picture": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "https://api.immopro.com/storage/profiles/1/photo.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fichier invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserResource": {
                "title": "Utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@email.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+22501234567",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "Côte d'Ivoire",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "example": "Abidjan",
                        "nullable": true
                    },
                    "profile_picture": {
                        "type": "string",
                        "format": "url",
                        "example": "https://...",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "example": "client",
                        "enum": [
                            "admin",
                            "agent",
                            "client"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "suspended",
                            "blocked"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu après connexion. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}